Add option to use distro provided certbot package
This patch allows a certbot package from a distro repository to be
used instead of the certbot-auto script. For ubuntu the distro package
takes care of setting up all the necessary directories in /etc/letsencrypt
and creates a systemd service for running the renewal. This avoids
needing to manage those aspects of the installation in this ansible role.

Change-Id: If7c5bcde299362bb3d2a136db659ca319e22e35b