Ensure there is no race between haproxy and certbot for initial cert
The certbot pre-hook is not used during initial setup of the cert,
only during renewal. This means that the same race condition exists
at initial configiuration as renewal. This patch uses the same
approach as used in the renewal pre-hook and applies it during
initialisation of certbot. This fixes race condition related failures
during initial provisioning of haproxy+letsencrypt.

Change-Id: Ica5ed5de24e3eb2fb5a743bb877d113ed0bb8a43